refactor: refacts `Instructions` component to be used in `forgot_pk | rejected` userStatus context
